+++ radarpipeline/features/__init__.py
@@ -1,3 +1,3 @@
 from radarpipeline.features.feature import Feature
 from radarpipeline.features.feature_group import FeatureGroup
-from radarpipeline.features.feature_loader import FeatureLoader
+#from radarpipeline.features.feature_loader import FeatureLoader
