+++ tests/tests_datalib/test_radar_user_data.py
@@ -1,9 +1,24 @@
 from radarpipeline.datalib import RadarVariableData, RadarUserData
-from radarpipeline.common.utils import PySparkTestCase
 import unittest
 import os
 import pandas as pd
-from pyspark.testing import assertDataFrameEqual
+from radarpipeline.project.sparkengine import SparkEngine
+import unittest
+from radarpipeline.common.utils import preprocess_time_data
+
+
+class PySparkTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.spark_engine = SparkEngine()
+        cls.spark = cls.spark_engine.initialize_spark_session()
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.spark_engine.close_spark_session()
+
+    def preprocess_data(self, data):
+        return preprocess_time_data(data)
 
 
 class TestRadarUserData(PySparkTestCase):
