+++ tests/tests_datalib/test_radardata.py
@@ -2,11 +2,26 @@ from radarpipeline.datalib import RadarVariableData, RadarUserData, RadarData
 import unittest
 import os
 import pandas as pd
-from radarpipeline.common.utils import PySparkTestCase
 import pyspark.sql.functions as f
 from pyspark.sql.types import TimestampType
 from pandas.testing import assert_frame_equal
+from radarpipeline.project.sparkengine import SparkEngine
+import unittest
+from radarpipeline.common.utils import preprocess_time_data
 
+
+class PySparkTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.spark_engine = SparkEngine()
+        cls.spark = cls.spark_engine.initialize_spark_session()
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.spark_engine.close_spark_session()
+
+    def preprocess_data(self, data):
+        return preprocess_time_data(data)
 
 class TestRadarData(PySparkTestCase):
