+++ tests/tests_features/test_featureloader.py
@@ -1,10 +1,10 @@
-from radarpipeline.features import FeatureLoader
-from pandas.testing import assert_frame_equal
-import os
-import pandas as pd
-import unittest
+#from radarpipeline.features import FeatureLoader
+#from pandas.testing import assert_frame_equal
+#import os
+#import pandas as pd
+#import unittest
 
-
+"""
 class TestFeatureLoader(unittest.TestCase):
     def setUp(self) -> None:
         self.PATH = "./mockdata/mockdata"
@@ -83,3 +83,4 @@ class TestFeatureLoader(unittest.TestCase):
                                actual_df.sort_values(['key.userId', 'value.time'])
                                .reset_index(drop=True), check_datetimelike_compat=True,
                                check_dtype=False)
+"""
