+++ tests/tests_io/test_reader.py
@@ -9,9 +9,24 @@ from radarpipeline.datalib import RadarVariableData, RadarUserData, RadarData
 from radarpipeline.common import constants
 from pandas.testing import assert_frame_equal
 from numpy.testing import assert_array_equal
-from radarpipeline.common.utils import PySparkTestCase
 from pyspark.sql.types import StructField, StructType
+from radarpipeline.project.sparkengine import SparkEngine
+import unittest
+from radarpipeline.common.utils import preprocess_time_data
 
+
+class PySparkTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.spark_engine = SparkEngine()
+        cls.spark = cls.spark_engine.initialize_spark_session()
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.spark_engine.close_spark_session()
+
+    def preprocess_data(self, data):
+        return preprocess_time_data(data)
 
 class TestSparkCSVDataReader(PySparkTestCase):
     def setUp(self):
